feat(domain): prevent deletion of primary addresses
Implemented a business rule that blocks deletion of addresses
when they are marked as primary for one or more users.
The rule is enforced consistently across Address and UserAddress
using a centralized domain exception and correct principal filtering.

+++ src/main/java/br/com/fiap/techchallenge/core/usecase/impl/useraddress/DeleteUserAddressUseCaseImpl.java
@@ -1,23 +1,46 @@
 package br.com.fiap.techchallenge.core.usecase.impl.useraddress;
 
+import br.com.fiap.techchallenge.core.domain.exception.useraddress.CannotDeletePrimaryAddressException;
 import br.com.fiap.techchallenge.core.domain.exception.useraddress.UserAddressNotFoundException;
 import br.com.fiap.techchallenge.core.domain.model.UserAddress;
 import br.com.fiap.techchallenge.core.usecase.in.useraddress.DeleteUserAddressUseCase;
+import br.com.fiap.techchallenge.core.usecase.out.AddressRepositoryPort;
 import br.com.fiap.techchallenge.core.usecase.out.UserAddressRepositoryPort;
+
+import java.util.List;
 
 public class DeleteUserAddressUseCaseImpl implements DeleteUserAddressUseCase {
 
     private final UserAddressRepositoryPort userAddressRepository;
+    private final AddressRepositoryPort addressRepository;
 
-    public DeleteUserAddressUseCaseImpl(UserAddressRepositoryPort userAddressRepository) {
+    public DeleteUserAddressUseCaseImpl(UserAddressRepositoryPort userAddressRepository, AddressRepositoryPort addressRepository) {
         this.userAddressRepository = userAddressRepository;
+        this.addressRepository = addressRepository;
     }
 
     @Override
-    public void execute(String id) {
-        UserAddress existing = userAddressRepository.findUserAddressById(id)
-                .orElseThrow(() -> new UserAddressNotFoundException(id));
+    public void execute(String userAddressId) {
 
+        // 1. Check if exists any link
+        UserAddress existing = userAddressRepository.findUserAddressById(userAddressId)
+                .orElseThrow(() -> new UserAddressNotFoundException(userAddressId));
+
+        //2. Find links with users
+        List<UserAddress> principalLinks =
+                userAddressRepository.findPrincipalsById(userAddressId);
+
+        // 3. Rule: cannot delete a primary address
+        if (!principalLinks.isEmpty()) {
+
+            List<String> userIds = principalLinks.stream()
+                    .map(UserAddress::getUserId)
+                    .toList();
+
+            throw new CannotDeletePrimaryAddressException(userAddressId, userIds);
+        }
+
+        //4. Delete UserAddress
         userAddressRepository.deleteById(existing.getId());
     }
 }
